Fix SonarCloud findings: extract auth helpers and named IP constant
- Move _create_and_login_admin/_create_and_login_user to DigiScriptTestCase
  base class, removing duplication between test_system_info and test_db_backups
- Extract hardcoded 8.8.8.8 to named _PROBE_HOST/_PROBE_PORT constants with
  a comment clarifying no data is sent
