+++ server/test/controllers/api/test_system_info.py
@@ -6,37 +6,6 @@ from test.conftest import DigiScriptTestCase
 
 
 class TestSystemInfoController(DigiScriptTestCase):
-    def _create_and_login_admin(self, username="admin", password="adminpass"):
-        self.fetch(
-            "/api/v1/auth/create",
-            method="POST",
-            body=escape.json_encode(
-                {"username": username, "password": password, "is_admin": True}
-            ),
-        )
-        resp = self.fetch(
-            "/api/v1/auth/login",
-            method="POST",
-            body=escape.json_encode({"username": username, "password": password}),
-        )
-        return escape.json_decode(resp.body)["access_token"]
-
-    def _create_and_login_user(self, admin_token, username="user", password="userpass"):
-        self.fetch(
-            "/api/v1/auth/create",
-            method="POST",
-            body=escape.json_encode(
-                {"username": username, "password": password, "is_admin": False}
-            ),
-            headers={"Authorization": f"Bearer {admin_token}"},
-        )
-        resp = self.fetch(
-            "/api/v1/auth/login",
-            method="POST",
-            body=escape.json_encode({"username": username, "password": password}),
-        )
-        return escape.json_decode(resp.body)["access_token"]
-
     def _fetch_system_info(self, token=None):
         headers = {}
         if token:
